auth0-poc/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "auth0-poc"
version = "0.1.0"
description = ""
authors = ["backwardspy <backwardspy@gmail.com>"]
readme = "README.md"
packages = [{include = "auth0_poc"}]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.3"
Flask-Cors = "^3.0.10"
pydantic = "^1.9.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.27.1"

[tool.poetry.group.dev.dependencies]
black = "^22.1.0"
pylint = "^2.12.2"
mypy = "^0.931"
pytest = "^7.0.1"
ptpython = "^3.0.20"
isort = "^5.10.1"

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
]
